some stuff

+++ src/Map/Coordinate.java
@@ -6,15 +6,15 @@ package Map;
  */
 public class Coordinate
 {
-    private int x;
-    private int y;
-    public Coordinate(int x, int y)
+    private double x;
+    private double y;
+    public Coordinate(double x, double y)
     {
         this.x=x;
         this.y=y;
     }
 
-    public int getX() {
+    public double getX() {
         return x;
     }
 
@@ -22,7 +22,7 @@ public class Coordinate
         this.x = x;
     }
 
-    public int getY() {
+    public double getY() {
         return y;
     }
